Moved sample data
